requirements.txt
Flask==3.0.3
stripe==10.12.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
mysql-connector-python==8.4.0
